+++ autode/species/molecule.py
@@ -66,9 +66,6 @@ class Molecule(Species):
         if smiles is not None:
             self._init_smiles(smiles)
 
-        elif atoms is not None:
-            make_graph(self)
-
         # If the name is unassigned use a more interpretable chemical formula
         if name == 'molecule' and self.atoms is not None:
             self.name = self.formula