+++ tests/test_molecule.py
@@ -18,10 +18,6 @@ here = os.path.dirname(os.path.abspath(__file__))
 
 def test_basic_attributes():
 
-    with pytest.raises(NoAtomsInMolecule):
-        Molecule(atoms=[])
-        Molecule()
-
     methane = Molecule(name='methane', smiles='C')
 
     assert methane.name == 'methane'
@@ -263,3 +259,9 @@ def test_defined_metal_spin_state():
 
     mol = Molecule(smiles='[Sc]C', mult=3)
     assert mol.mult == 3
+
+
+def test_atom_class_defined_for_organic():
+
+    mol = Molecule(smiles='[Br-:1]')
+    assert mol.atoms[0].atom_class is not None