+++ doc/changelog.rst
@@ -2,6 +2,31 @@ Changelog
 =========
 
 
+1.2.3
+--------
+----------
+
+Minor functionality improvements and bugfixes.
+
+
+Usability improvements/Changes
+******************************
+- All exceptions now inherit from a base :code:`autode.exceptions.AutodeException`
+- Fixes a typo in :code:`autode.exceptions.UnsupoportedCalculationInput`
+- Adds documentation explaining the intention of each exception in  :code:`autode.exceptions`
+- Molecular graphs are now 'laziliy-loaded' i.e. generated once when the property is accessed
+
+
+Functionality improvements
+**************************
+- Adds the ability to define atom classes for molecules in turn allowing for identity reactions to be calculated
+
+
+Bug Fixes
+*********
+- Fixes clashing names for a reaction initialised explicitly from molecules without defined names
+
+
 1.2.2
 --------
 ----------